添加了打分脚本，调整目录

+++ lab_advanced/lab4/test.sh
@@ -1,7 +1,11 @@
 size=100
 
-rm -r ./input
-rm -r ./output
+if [ -e "input" ]; then
+    rm -r ./input
+fi
+if [ -e "output" ]; then
+    rm -r ./output
+fi
 
 python3 ./data_gen.py $size
